+++ sdk/managementgroups/azure-mgmt-managementgroups/azure/mgmt/managementgroups/aio/operations/_management_group_subscriptions_operations.py
@@ -1,4 +1,4 @@
-# pylint: disable=too-many-lines
+# pylint: disable=too-many-lines,too-many-statements
 # coding=utf-8
 # --------------------------------------------------------------------------
 # Copyright (c) Microsoft Corporation. All rights reserved.
@@ -7,7 +7,7 @@
 # Changes may cause incorrect behavior and will be lost if the code is regenerated.
 # --------------------------------------------------------------------------
 import sys
-from typing import Any, AsyncIterable, Callable, Dict, Optional, TypeVar
+from typing import Any, AsyncIterable, Callable, Dict, Optional, Type, TypeVar
 import urllib.parse
 
 from azure.core.async_paging import AsyncItemPaged, AsyncList
@@ -20,27 +20,24 @@ from azure.core.exceptions import (
     map_error,
 )
 from azure.core.pipeline import PipelineResponse
-from azure.core.pipeline.transport import AsyncHttpResponse
-from azure.core.rest import HttpRequest
+from azure.core.rest import AsyncHttpResponse, HttpRequest
 from azure.core.tracing.decorator import distributed_trace
 from azure.core.tracing.decorator_async import distributed_trace_async
 from azure.core.utils import case_insensitive_dict
 from azure.mgmt.core.exceptions import ARMErrorFormat
 
 from ... import models as _models
-from ..._vendor import _convert_request
 from ...operations._management_group_subscriptions_operations import (
     build_create_request,
     build_delete_request,
     build_get_subscription_request,
     build_get_subscriptions_under_management_group_request,
 )
-from .._vendor import ManagementGroupsAPIMixinABC
 
-if sys.version_info >= (3, 8):
-    from typing import Literal  # pylint: disable=no-name-in-module, ungrouped-imports
+if sys.version_info >= (3, 9):
+    from collections.abc import MutableMapping
 else:
-    from typing_extensions import Literal  # type: ignore  # pylint: disable=ungrouped-imports
+    from typing import MutableMapping  # type: ignore  # pylint: disable=ungrouped-imports
 T = TypeVar("T")
 ClsType = Optional[Callable[[PipelineResponse[HttpRequest, AsyncHttpResponse], T, Dict[str, Any]], Any]]
 
@@ -77,12 +74,11 @@ class ManagementGroupSubscriptionsOperations:
         :param cache_control: Indicates whether the request should utilize any caches. Populate the
          header with 'no-cache' value to bypass existing caches. Default value is "no-cache".
         :type cache_control: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
         :return: SubscriptionUnderManagementGroup or the result of cls(response)
         :rtype: ~azure.mgmt.managementgroups.models.SubscriptionUnderManagementGroup
         :raises ~azure.core.exceptions.HttpResponseError:
         """
-        error_map = {
+        error_map: MutableMapping[int, Type[HttpResponseError]] = {
             401: ClientAuthenticationError,
             404: ResourceNotFoundError,
             409: ResourceExistsError,
@@ -93,25 +89,22 @@ class ManagementGroupSubscriptionsOperations:
         _headers = kwargs.pop("headers", {}) or {}
         _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
-        api_version = kwargs.pop(
-            "api_version", _params.pop("api-version", self._config.api_version)
-        )  # type: Literal["2021-04-01"]
-        cls = kwargs.pop("cls", None)  # type: ClsType[_models.SubscriptionUnderManagementGroup]
+        api_version: str = kwargs.pop("api_version", _params.pop("api-version", self._config.api_version))
+        cls: ClsType[_models.SubscriptionUnderManagementGroup] = kwargs.pop("cls", None)
 
-        request = build_create_request(
+        _request = build_create_request(
             group_id=group_id,
             subscription_id=subscription_id,
             cache_control=cache_control,
             api_version=api_version,
-            template_url=self.create.metadata["url"],
             headers=_headers,
             params=_params,
         )
-        request = _convert_request(request)
-        request.url = self._client.format_url(request.url)  # type: ignore
+        _request.url = self._client.format_url(_request.url)
 
-        pipeline_response = await self._client._pipeline.run(  # type: ignore # pylint: disable=protected-access
-            request, stream=False, **kwargs
+        _stream = False
+        pipeline_response: PipelineResponse = await self._client._pipeline.run(  # pylint: disable=protected-access
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
@@ -121,14 +114,12 @@ class ManagementGroupSubscriptionsOperations:
             error = self._deserialize.failsafe_deserialize(_models.ErrorResponse, pipeline_response)
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
-        deserialized = self._deserialize("SubscriptionUnderManagementGroup", pipeline_response)
+        deserialized = self._deserialize("SubscriptionUnderManagementGroup", pipeline_response.http_response)
 
         if cls:
-            return cls(pipeline_response, deserialized, {})
+            return cls(pipeline_response, deserialized, {})  # type: ignore
 
-        return deserialized
-
-    create.metadata = {"url": "/providers/Microsoft.Management/managementGroups/{groupId}/subscriptions/{subscriptionId}"}  # type: ignore
+        return deserialized  # type: ignore
 
     @distributed_trace_async
     async def delete(  # pylint: disable=inconsistent-return-statements
@@ -143,12 +134,11 @@ class ManagementGroupSubscriptionsOperations:
         :param cache_control: Indicates whether the request should utilize any caches. Populate the
          header with 'no-cache' value to bypass existing caches. Default value is "no-cache".
         :type cache_control: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
         :return: None or the result of cls(response)
         :rtype: None
         :raises ~azure.core.exceptions.HttpResponseError:
         """
-        error_map = {
+        error_map: MutableMapping[int, Type[HttpResponseError]] = {
             401: ClientAuthenticationError,
             404: ResourceNotFoundError,
             409: ResourceExistsError,
@@ -159,25 +149,22 @@ class ManagementGroupSubscriptionsOperations:
         _headers = kwargs.pop("headers", {}) or {}
         _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
-        api_version = kwargs.pop(
-            "api_version", _params.pop("api-version", self._config.api_version)
-        )  # type: Literal["2021-04-01"]
-        cls = kwargs.pop("cls", None)  # type: ClsType[None]
+        api_version: str = kwargs.pop("api_version", _params.pop("api-version", self._config.api_version))
+        cls: ClsType[None] = kwargs.pop("cls", None)
 
-        request = build_delete_request(
+        _request = build_delete_request(
             group_id=group_id,
             subscription_id=subscription_id,
             cache_control=cache_control,
             api_version=api_version,
-            template_url=self.delete.metadata["url"],
             headers=_headers,
             params=_params,
         )
-        request = _convert_request(request)
-        request.url = self._client.format_url(request.url)  # type: ignore
+        _request.url = self._client.format_url(_request.url)
 
-        pipeline_response = await self._client._pipeline.run(  # type: ignore # pylint: disable=protected-access
-            request, stream=False, **kwargs
+        _stream = False
+        pipeline_response: PipelineResponse = await self._client._pipeline.run(  # pylint: disable=protected-access
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
@@ -188,9 +175,7 @@ class ManagementGroupSubscriptionsOperations:
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
         if cls:
-            return cls(pipeline_response, None, {})
-
-    delete.metadata = {"url": "/providers/Microsoft.Management/managementGroups/{groupId}/subscriptions/{subscriptionId}"}  # type: ignore
+            return cls(pipeline_response, None, {})  # type: ignore
 
     @distributed_trace_async
     async def get_subscription(
@@ -205,12 +190,11 @@ class ManagementGroupSubscriptionsOperations:
         :param cache_control: Indicates whether the request should utilize any caches. Populate the
          header with 'no-cache' value to bypass existing caches. Default value is "no-cache".
         :type cache_control: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
         :return: SubscriptionUnderManagementGroup or the result of cls(response)
         :rtype: ~azure.mgmt.managementgroups.models.SubscriptionUnderManagementGroup
         :raises ~azure.core.exceptions.HttpResponseError:
         """
-        error_map = {
+        error_map: MutableMapping[int, Type[HttpResponseError]] = {
             401: ClientAuthenticationError,
             404: ResourceNotFoundError,
             409: ResourceExistsError,
@@ -221,25 +205,22 @@ class ManagementGroupSubscriptionsOperations:
         _headers = kwargs.pop("headers", {}) or {}
         _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
-        api_version = kwargs.pop(
-            "api_version", _params.pop("api-version", self._config.api_version)
-        )  # type: Literal["2021-04-01"]
-        cls = kwargs.pop("cls", None)  # type: ClsType[_models.SubscriptionUnderManagementGroup]
+        api_version: str = kwargs.pop("api_version", _params.pop("api-version", self._config.api_version))
+        cls: ClsType[_models.SubscriptionUnderManagementGroup] = kwargs.pop("cls", None)
 
-        request = build_get_subscription_request(
+        _request = build_get_subscription_request(
             group_id=group_id,
             subscription_id=subscription_id,
             cache_control=cache_control,
             api_version=api_version,
-            template_url=self.get_subscription.metadata["url"],
             headers=_headers,
             params=_params,
         )
-        request = _convert_request(request)
-        request.url = self._client.format_url(request.url)  # type: ignore
+        _request.url = self._client.format_url(_request.url)
 
-        pipeline_response = await self._client._pipeline.run(  # type: ignore # pylint: disable=protected-access
-            request, stream=False, **kwargs
+        _stream = False
+        pipeline_response: PipelineResponse = await self._client._pipeline.run(  # pylint: disable=protected-access
+            _request, stream=_stream, **kwargs
         )
 
         response = pipeline_response.http_response
@@ -249,14 +230,12 @@ class ManagementGroupSubscriptionsOperations:
             error = self._deserialize.failsafe_deserialize(_models.ErrorResponse, pipeline_response)
             raise HttpResponseError(response=response, model=error, error_format=ARMErrorFormat)
 
-        deserialized = self._deserialize("SubscriptionUnderManagementGroup", pipeline_response)
+        deserialized = self._deserialize("SubscriptionUnderManagementGroup", pipeline_response.http_response)
 
         if cls:
-            return cls(pipeline_response, deserialized, {})
+            return cls(pipeline_response, deserialized, {})  # type: ignore
 
-        return deserialized
-
-    get_subscription.metadata = {"url": "/providers/Microsoft.Management/managementGroups/{groupId}/subscriptions/{subscriptionId}"}  # type: ignore
+        return deserialized  # type: ignore
 
     @distributed_trace
     def get_subscriptions_under_management_group(
@@ -272,7 +251,6 @@ class ManagementGroupSubscriptionsOperations:
          include a token parameter that specifies a starting point to use for subsequent calls. Default
          value is None.
         :type skiptoken: str
-        :keyword callable cls: A custom type or function that will be passed the direct response
         :return: An iterator like instance of either SubscriptionUnderManagementGroup or the result of
          cls(response)
         :rtype:
@@ -282,12 +260,10 @@ class ManagementGroupSubscriptionsOperations:
         _headers = kwargs.pop("headers", {}) or {}
         _params = case_insensitive_dict(kwargs.pop("params", {}) or {})
 
-        api_version = kwargs.pop(
-            "api_version", _params.pop("api-version", self._config.api_version)
-        )  # type: Literal["2021-04-01"]
-        cls = kwargs.pop("cls", None)  # type: ClsType[_models.ListSubscriptionUnderManagementGroup]
+        api_version: str = kwargs.pop("api_version", _params.pop("api-version", self._config.api_version))
+        cls: ClsType[_models.ListSubscriptionUnderManagementGroup] = kwargs.pop("cls", None)
 
-        error_map = {
+        error_map: MutableMapping[int, Type[HttpResponseError]] = {
             401: ClientAuthenticationError,
             404: ResourceNotFoundError,
             409: ResourceExistsError,
@@ -298,16 +274,14 @@ class ManagementGroupSubscriptionsOperations:
         def prepare_request(next_link=None):
             if not next_link:
 
-                request = build_get_subscriptions_under_management_group_request(
+                _request = build_get_subscriptions_under_management_group_request(
                     group_id=group_id,
                     skiptoken=skiptoken,
                     api_version=api_version,
-                    template_url=self.get_subscriptions_under_management_group.metadata["url"],
                     headers=_headers,
                     params=_params,
                 )
-                request = _convert_request(request)
-                request.url = self._client.format_url(request.url)  # type: ignore
+                _request.url = self._client.format_url(_request.url)
 
             else:
                 # make call to next link with the client's api-version
@@ -319,26 +293,26 @@ class ManagementGroupSubscriptionsOperations:
                     }
                 )
                 _next_request_params["api-version"] = self._config.api_version
-                request = HttpRequest(
+                _request = HttpRequest(
                     "GET", urllib.parse.urljoin(next_link, _parsed_next_link.path), params=_next_request_params
                 )
-                request = _convert_request(request)
-                request.url = self._client.format_url(request.url)  # type: ignore
-                request.method = "GET"
-            return request
+                _request.url = self._client.format_url(_request.url)
+                _request.method = "GET"
+            return _request
 
         async def extract_data(pipeline_response):
             deserialized = self._deserialize("ListSubscriptionUnderManagementGroup", pipeline_response)
             list_of_elem = deserialized.value
             if cls:
-                list_of_elem = cls(list_of_elem)
+                list_of_elem = cls(list_of_elem)  # type: ignore
             return deserialized.next_link or None, AsyncList(list_of_elem)
 
         async def get_next(next_link=None):
-            request = prepare_request(next_link)
+            _request = prepare_request(next_link)
 
-            pipeline_response = await self._client._pipeline.run(  # type: ignore # pylint: disable=protected-access
-                request, stream=False, **kwargs
+            _stream = False
+            pipeline_response: PipelineResponse = await self._client._pipeline.run(  # pylint: disable=protected-access
+                _request, stream=_stream, **kwargs
             )
             response = pipeline_response.http_response
 
@@ -350,5 +324,3 @@ class ManagementGroupSubscriptionsOperations:
             return pipeline_response
 
         return AsyncItemPaged(get_next, extract_data)
-
-    get_subscriptions_under_management_group.metadata = {"url": "/providers/Microsoft.Management/managementGroups/{groupId}/subscriptions"}  # type: ignore